requirements-dev.txt
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0
pytest-mock>=3.0
ruff>=0.1
mypy>=1.0
types-PyYAML>=6.0
types-requests>=2.31
pandas-stubs>=2.0
types-beautifulsoup4>=4.12
